requirements.txt
python-telegram-bot==22.0
psycopg2-binary==2.9.10